Move aligned_allocator to vtr_memory.
